requirements.txt
requests
pandas
pyarrow
python-dotenv
duckdb==0.10.2
streamlit